m(poster): drop Carl Hewitt reference
I instructed Hannah to add a citation when talking about the Actor
Model, but, it seems that it isn't rendered well. Given that (1) we
don't have space left in the poster for a references box, (2) this is
the only reference, and (3) the claim of "Teoria matemática
originalmente concebida para sistemas distribuídos" is acceptable to
make without references, drop this Carl Hewitt reference.

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>